requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
